Move function select_templated_tests
Move function `select_templated_tests` to `ssg` module so that
it can be reused outside of `ssg_test_suite` module.
